Improve test coverage.
